Refactored my channel endpoints to remove distinguishment between your channel and other channels
