+++ Tribler/Core/Modules/restapi/channels/__init__.py
@@ -1,0 +1,3 @@
+"""
+This module contains endpoints to manage everything in a single channel such as torrents, rss feeds and playlists.
+"""
