+++ Tribler/Core/Modules/restapi/root_endpoint.py
@@ -1,6 +1,6 @@
 from twisted.web import resource
+from Tribler.Core.Modules.restapi.channels.channels_endpoint import ChannelsEndpoint
 
-from Tribler.Core.Modules.restapi.channels_endpoint import ChannelsEndpoint
 from Tribler.Core.Modules.restapi.downloads_endpoint import DownloadsEndpoint
 from Tribler.Core.Modules.restapi.events_endpoint import EventsEndpoint
 from Tribler.Core.Modules.restapi.my_channel_endpoint import MyChannelEndpoint
@@ -19,8 +19,9 @@ class RootEndpoint(resource.Resource):
         resource.Resource.__init__(self)
         self.session = session
 
-        child_handler_dict = {"search": SearchEndpoint, "channels": ChannelsEndpoint, "mychannel": MyChannelEndpoint,
-                              "settings": SettingsEndpoint, "variables": VariablesEndpoint, "events": EventsEndpoint,
+        child_handler_dict = {"search": SearchEndpoint, "channels": ChannelsEndpoint,
+                              "mychannel": MyChannelEndpoint, "settings": SettingsEndpoint,
+                              "variables": VariablesEndpoint, "events": EventsEndpoint,
                               "downloads": DownloadsEndpoint}
 
         for path, child_cls in child_handler_dict.iteritems():
